requirements.txt
audio_augmentations==0.1.3
avalanche_lib==0.3.1
finetuning_scheduler==2.0.2
Flask==2.2.3
h5py==3.7.0
kymatio==0.3.0
librosa==0.10.0.post2
lightning==2.0.1.post0
matplotlib==3.7.1
nemo_toolkit==1.16.0
numpy==1.24.3
pandas==2.0.0
pymongo==3.12.0
sacred==0.8.4
seaborn==0.12.2
sounddevice==0.4.6
tensorflow==2.11.0
torch==1.13.1
torch_audiomentations==0.11.0
torchaudio==0.13.1
torchinfo==1.7.2
torchmetrics==0.11.4
tqdm==4.65.0
transformers==4.27.2
